+++ tests/integration/api/test-crud-functions.php
@@ -6,8 +6,8 @@ use WPE\AtlasContentModeler\ContentConnect\Relationships\PostToPost;
 use function WPE\AtlasContentModeler\API\add_relationship;
 use function WPE\AtlasContentModeler\API\get_relationship;
 use function WPE\AtlasContentModeler\API\replace_relationship;
-use function WPE\AtlasContentModeler\API\fetch_model;
-use function WPE\AtlasContentModeler\API\fetch_model_field;
+use function WPE\AtlasContentModeler\API\get_model;
+use function WPE\AtlasContentModeler\API\get_model_field;
 use function WPE\AtlasContentModeler\API\insert_model_entry;
 use function WPE\AtlasContentModeler\API\update_model_entry;
 
@@ -138,16 +138,16 @@ class TestApiFunctions extends Integration_TestCase {
 		$this->assertEquals( [ 'The post ID 4444444 was not found' ], $updated_id->get_error_messages( 'model_entry_not_found' ) );
 	}
 
-	public function test_fetch_model_returns_the_model_schema_if_exists() {
-		$this->assertEquals( $this->content_models['person'], fetch_model( 'person' ) );
+	public function test_get_model_returns_the_model_schema_if_exists() {
+		$this->assertEquals( $this->content_models['person'], get_model( 'person' ) );
 	}
 
-	public function test_fetch_model_returns_null_if_the_model_does_not_exist() {
-		$this->assertNull( fetch_model( 'does_not_exist' ) );
+	public function test_get_model_returns_null_if_the_model_does_not_exist() {
+		$this->assertNull( get_model( 'does_not_exist' ) );
 	}
 
-	public function test_fetch_model_field_returns_the_field_if_exists() {
-		$model_field = fetch_model_field( 'person', 'name' );
+	public function test_get_model_field_returns_the_field_if_exists() {
+		$model_field = get_model_field( 'person', 'name' );
 
 		$this->assertEquals(
 			$this->content_models['person']['fields']['1648575961490'],
@@ -155,14 +155,14 @@ class TestApiFunctions extends Integration_TestCase {
 		);
 	}
 
-	public function test_fetch_model_field_returns_null_if_the_field_does_not_exist() {
-		$model_field = fetch_model_field( 'person', 'does_not_exist' );
+	public function test_get_model_field_returns_null_if_the_field_does_not_exist() {
+		$model_field = get_model_field( 'person', 'does_not_exist' );
 
 		$this->assertNull( $model_field );
 	}
 
-	public function test_fetch_model_field_returns_null_if_the_model_does_not_exist() {
-		$model_field = fetch_model_field( 'does_not_exist', 'name' );
+	public function test_get_model_field_returns_null_if_the_model_does_not_exist() {
+		$model_field = get_model_field( 'does_not_exist', 'name' );
 
 		$this->assertNull( $model_field );
 	}