+++ CHANGELOG.md
@@ -11,8 +11,8 @@
 - Function replace_relationship() to replace/add model entry relationships.
 - Function add_relationship() to append model entry relationships.
 - Function get_relationship() to retrieve the relationship object.
-- Function fetch_model() to retrieve the model schema.
-- Function fetch_model_field() to retrieve the model field schema.
+- Function get_model() to retrieve the model schema.
+- Function get_model_field() to retrieve the model field schema.
 
 ## 0.16.0 - 2022-04-06
 ### Added